+++ R/updateCorrespondenceTable.R
@@ -140,20 +140,38 @@
 #' library. All sample datasets may be found there.
 #' }
 #' @examples
-#'  \dontrun{
-#'  ## Application of function updateCorrespondenceTable() with CN 2021 being the
-#'  ## original classification A, CPA 2.1 being the target classification B, CN 2022
-#'  ## being the updated version A*, CN 2021:CPA 2.1 being the previous correspondence
-#'  ## table A:B, and CN 2021:CN 2022 being the A:A* concordance table. The desired
+#'  {
+#'  ## Application of function updateCorrespondenceTable() with NAICS 2017 being the
+#'  ## original classification A, NACE being the target classification B, NAICS 2022
+#'  ## being the updated version A*, NAICS 2017:NACE being the previous correspondence
+#'  ## table A:B, and NAICS 2017:NAICS 2022 being the A:A* concordance table. The desired
 #'  ## name for the csv file that will contain the updated correspondence table is
-#'  ## "updateCorrespondenceTable.csv", the reference classification is B, and the
+#'  ## "updateCorrespondenceTable.csv", there is no reference classification, and the
 #'  ## maximum acceptable proportions of unmatched codes between the original
 #'  ## classification A and the target classification B, and between the original
-#'  ## classification A and the updated classification A* are 0.4 and 0.4, respectively.
+#'  ## classification A and the updated classification A* are 0.5 and 0.3, respectively.
 #'     
-#'  UPC <- updateCorrespondenceTable("CN2021.csv", "CPA21.csv", "CN2022.csv",
-#'                                   "CN2021_CPA21.csv", "CN2021_CN2022.csv", 
-#'                                   "updateCorrespondenceTable.csv", "B", 0.4, 0.4)
+#'  A <- system.file("extdata", "NAICS2017.csv", package = "correspondenceTables")
+#'  AStar <- system.file("extdata", "NAICS2022.csv", package = "correspondenceTables")
+#'  B <- system.file("extdata", "NACE.csv", package = "correspondenceTables")
+#'  AB <- system.file("extdata", "NAICS2017_NACE.csv", package = "correspondenceTables")
+#'  AAStar <- system.file("extdata", "NAICS2017_NAICS2022.csv", package = "correspondenceTables")
+#'  
+#'  UPC <- updateCorrespondenceTable(A,
+#'                                   B,
+#'                                   AStar, 
+#'                                   AB, 
+#'                                   AAStar, 
+#'                                   "updateCorrespondenceTable.csv", 
+#'                                   "none", 
+#'                                   0.5, 
+#'                                   0.3)
+#'  
+#'  summary(UPC)
+#'  head(UPC$updateCorrespondenceTable)
+#'  UPC$classificationNames
+#'  unlink("updateCorrespondenceTable.csv")
+#'  unlink("classificationNames_updateCorrespondenceTable.csv")
 #'     }
 
 
@@ -312,13 +330,13 @@ updateCorrespondenceTable <- function(A, B, AStar, AB, AAStar, CSVout = NULL, Re
     }
 
     if (sum(!is.na(match(classA[, 1], corrAAStar[, 1]))) == 0) {
-        cat(paste("WARNING: there is no code of ", colnames(classA)[1], " that appears in both ",
+        message(paste("WARNING: there is no code of ", colnames(classA)[1], " that appears in both ",
             test.names[1], " and ", test.names[5], ". When the execution of the function is over, please check the files to ensure that this is not the result of a mistake in their preparation or declaration.",
             sep = ""))
     }
 
     if (sum(!is.na(match(classA[, 1], corrAB[, 1]))) == 0) {
-        cat(paste("WARNING: there is no code of ", colnames(classA)[1], " that appears in both ",
+        message(paste("WARNING: there is no code of ", colnames(classA)[1], " that appears in both ",
             test.names[1], " and ", test.names[4], ". When the execution of the function is over, please check the files to ensure that this is not the result of a mistake in their preparation or declaration.",
             sep = ""))
     }