+++ R/zzz.R
@@ -1,5 +1,5 @@
 #' @importFrom readr write_excel_csv
 #' @importFrom stats aggregate
-#' @importFrom utils menu read.csv tail write.csv
+#' @importFrom utils menu read.csv tail write.csv txtProgressBar setTxtProgressBar
  
 .onLoad<-function(libname, pkgname){}