+++ NEWS.md
@@ -1,7 +1,13 @@
+# correspondenceTables 0.7.3
+
+- aligning to the comments from CRAN 
+- using 'txtProgressBar()' and 'setTxtProgressBar()' functions
+- replacing 'cat()' with 'message()' function
+- examples updated to be able to test it automatically 
+
 # correspondenceTables 0.7.2
 
 - adjustment for CRAN approval
-
 
 # correspondenceTables 0.7.1